% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeontario-shape.R
\name{base_ontarioshore}
\alias{base_ontarioshore}
\title{base_ontarioshore}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Ontario as base ggplot object
}
\examples{
\dontrun{
data("base_ontarioshore")
base_ontarioshore
}

}
\seealso{
shape_ontarioshore
}
