% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2coorddd.R
\name{convert2coorddd}
\alias{convert2coorddd}
\title{convert2coorddd}
\usage{
convert2coorddd(coord)
}
\arguments{
\item{coord}{A coordinate in the for DDMM.ddd}
}
\description{
Parses coordinates from a NMEA string to Deg_DD
}
\examples{
convert2coorddd(4351.986)
}
