% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEED_functions.R
\name{feed_convert_coords}
\alias{feed_convert_coords}
\title{feed_convert_coords}
\usage{
feed_convert_coords(feedcoord)
}
\arguments{
\item{feedcoord}{a lat or long in the FEEd format of DDMM.mmm}
}
\value{
coordinate value as DD.ddd
}
\description{
FEED stores coordinate as `DDMM.ddd`. This function converts
that format to `DD.ddd` for plotting purposes. Note: Longitude will normally require
the value to be multiplied by `-1 ` for sites in the western hemisphere
}
\examples{
feed_convert_coords(7705.819) * (-1)
feed_convert_coords(4402.984)
}
