% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRBR.R
\name{import_foot_rope_files}
\alias{import_foot_rope_files}
\title{import_footrope_files}
\usage{
import_foot_rope_files(rbrfiles)
}
\arguments{
\item{rskfile}{a *.rsk file from an RBR temp/depth logger that has been parsed and appended with 'foot'}
}
\description{
A wrapper for `importRBR` specific to importing footrope files
}
