% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRBR.R
\name{import_merge_tow_files}
\alias{import_merge_tow_files}
\title{import_merge_tow_files}
\usage{
import_merge_tow_files(rbrfiles)
}
\arguments{
\item{rskfile}{a *.rsk file from an RBR temp/depth logger that has been parsed and appended with 'head' or 'foot'}
}
\description{
A wrapper for `importRBR` specific to merging head and footrope files
}
