% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRBR.R
\name{importRBR}
\alias{importRBR}
\title{importRBR}
\usage{
importRBR(rskfile)
}
\arguments{
\item{rskfile}{a *.rsk file from an RBR temp/depth logger}
}
\description{
A wrapper for the `oce` function `read.rsk`. This wrapper converts the raw pressure data
and converts it to depth in freshwater.
}
