% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{LObath100m}
\alias{LObath100m}
\title{Lake Ontario 100m Bathymetry Shapefile}
\format{
An object of class \code{SpatialLinesDataFrame} with 65 rows and 10 columns.
}
\usage{
LObath100m
}
\description{
A shapefile that has been imported using `readORG` from rgdal
to create a spatial data file that can be plotted.
Data is a SpatialPolygonsDataFrame [package "sp"]
}
\keyword{datasets}
