% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{LOschaner}
\alias{LOschaner}
\title{Schaner Outline}
\format{
A data frame with 1646 observations and 4 variables
}
\usage{
LOschaner
}
\description{
A dataset that contains coordinate data to create an outline
and bathymetric contours for Lake Ontario
}
\keyword{datasets}
