% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{LOshore}
\alias{LOshore}
\title{Lake Ontario Shoreline Shapefile}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 2 columns.
}
\usage{
LOshore
}
\description{
A shapefile that has been imported using `readORG` from rgdal
to create a spatial data file that can be plotted.
Data is a SpatialPolygonsDataFrame [package "sp"]
}
\keyword{datasets}
