% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapFunctions.R
\name{LOspatial}
\alias{LOspatial}
\title{Lake Ontario Spatial Map}
\usage{
LOspatial()
}
\description{
This implemetation of the Lake Ontario base map utilizes
shape files uploaded within this package. This function requires
that the package `sp` is installed.
}
\seealso{
plot(LOshore)
}
