% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_bottom_contact.R
\name{pick_bottom_contact_from_foot}
\alias{pick_bottom_contact_from_foot}
\title{Pick bottom contact time from footrope only}
\usage{
pick_bottom_contact_from_foot(onetow)
}
\arguments{
\item{onetow}{expects an RBR temperature profiles}
}
\value{
dataframe with start, end and measured bottom contact time (in seconds) for each tow provided
}
\description{
A function that first creates a plot then provides a cursor for the user to identify the start and end of where
the trawl contacted bottom (or in the case of a midwater trawl, reached the target fishing depth).
}
\examples{
\dontrun{
library(lakeontario)
rbrfiles <- dir("Data/RBRClean", full.names = T)
towstats <- import_foot_rope_files(rbrfiles)
head(towstats)
str(towstats)
tow_stat_list <- split(towstats, towstats$SERIAL)
lapply(onetow, pick_bottom_contact_from_foot)
}

}
\seealso{
[rbr_tow_plot()]

[pick_bottom_contact()]
}
