% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_setup.R
\name{proj_setup}
\alias{proj_setup}
\title{proj_setup}
\usage{
proj_setup(path, ...)
}
\arguments{
\item{path}{file path that is populated during the setup procedure}

\item{...}{additional arguments as required}
}
\description{
A project setup function that is used for new RStudio project template
}
