% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRBR.R
\name{rbr_tow_plot}
\alias{rbr_tow_plot}
\title{rbr_tow_plot}
\usage{
rbr_tow_plot(oneserial)
}
\arguments{
\item{oneserial}{a dataframe that contains depth and temp info from a single tow}
}
\description{
plots a tow profile
}
