% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk_shift_time.R
\name{rsk_shift_time}
\alias{rsk_shift_time}
\title{Shift rsk time stamp}
\usage{
rsk_shift_time(rskfile, shiftdays)
}
\arguments{
\item{rskfile}{a path to an rsk file (RBR depth/temp logger)}

\item{shiftdays}{amount of time, in days, to adjust the timestamps}
}
\value{
saves modified tables to the original rsk file
}
\description{
An rsk file used with RBR loggers is a specialized SQLite database.
Given this data structure it is possible to use R to connect to the raw data, manipulate
it to fix a time stamp that was recorded improperly as a result of a reset of the
internal clock. The function takes an assigned time shift, in days, and applies it
to the raw SQLite data such that the RSK file can be read in Ruskin with the correct
time stamp.
}
\examples{
\dontrun{
rsk_shift_time("RBRraw/tow1.rsk", shiftdays = 0.5)
rsk_shift_time("RBRraw/tow1.rsk", shiftdays = 272)
}
}
