% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{RV2FN_SPECIES}
\alias{RV2FN_SPECIES}
\title{RV2FN_SPECIES}
\format{
A data frame
}
\usage{
RV2FN_SPECIES
}
\description{
RVCAT reference table for SPECIES code and corresponding SPC code (FN2)
}
\keyword{datasets}
