% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapFunctions.R
\name{SchanerMap}
\alias{SchanerMap}
\title{Lake Ontario Base Map}
\usage{
SchanerMap()
}
\description{
Ted Schaner created a simple csv file that contains the Lake Ontario shoreline
and the 50, 100 and 150m bathymetric contours. This function plots the base map upon which
additional data layers can be plotted.
}
