% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeontario-shape.R
\name{shape_ontarioshore}
\alias{shape_ontarioshore}
\title{shape_ontarioshore}
\format{
an sf spatial data frame
}
\description{
Lake Ontario as shapefile
}
\examples{
\dontrun{
data("shape_ontarioshore")
class(shape_ontarioshore)
plot(shape_ontarioshore)
}
}
\seealso{
base_ontarioshore
}
