% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOC2deg.R
\name{SILOCgetLAT10}
\alias{SILOCgetLAT10}
\title{Parses LAT from a 10 character SILOC}
\usage{
SILOCgetLAT10(fn121column)
}
\arguments{
\item{fn121column}{specify the FN2 column name}
}
\description{
Helper function for SILOC2COORD to get LAT
}
