% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOC2deg.R
\name{SILOCgetLAT12}
\alias{SILOCgetLAT12}
\title{Parses LAT from a 12 character SILOC}
\usage{
SILOCgetLAT12(fn121column)
}
\arguments{
\item{fn121column}{specify the FN2 column name}
}
\description{
Helper function for SILOC2COORD to get LAT
}
