% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOC2deg.R
\name{SILOCgetLON10}
\alias{SILOCgetLON10}
\title{Parses LON from a 10 characterSILOC}
\usage{
SILOCgetLON10(fn121column)
}
\arguments{
\item{fn121column}{specify the FN2 column name}
}
\description{
Helper function for SILOC2COORD to get LON
}
