% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOC2deg.R
\name{SILOCgetLON12}
\alias{SILOCgetLON12}
\title{Parses LON from a 12 character SILOC}
\usage{
SILOCgetLON12(fn121column)
}
\arguments{
\item{fn121column}{specify the FN2 column name}
}
\description{
Helper function for SILOC2COORD to get LON
}
