% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotThemes.R
\name{theme_gfs}
\alias{theme_gfs}
\title{LOMU AR ggplot2 Theme}
\usage{
theme_gfs(base_size = 17, base_family = "sans")
}
\arguments{
\item{base_size}{Default font size value of 17}

\item{base_family}{Default font family is sans}
}
\description{
This theme provides an figure layout that is
consistent with the style guidelines for the LOMU Annual Report.
}
\examples{
require(ggplot2)
qplot(1:10, 1:10) + theme_gfs()
}
\author{
Jeremy Holden and Ron Green
}
