% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotThemes.R
\name{theme_gfs2}
\alias{theme_gfs2}
\title{Another LOMU AR ggplot2 Theme}
\usage{
theme_gfs2(base_size = 17, base_family = "sans")
}
\arguments{
\item{base_size}{Default font size value of 17}

\item{base_family}{Default font family is sans}
}
\description{
This theme provides an figure layout that is
consistent with the style guidelines for the LOMU Annual Report
that slight, but possibly important changes.
}
\examples{
require(ggplot2)
qplot(1:10, 1:10, geom = "point") + theme_gfs2()
}
\author{
Jeremy Holden and Ron Green
}
