% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotThemes.R
\name{theme_hacmap}
\alias{theme_hacmap}
\title{LOMU AR ggplot2 Acoustic Map Theme}
\usage{
theme_hacmap(base_size = 17, base_family = "sans")
}
\arguments{
\item{base_size}{Default font size value of 17}

\item{base_family}{Default font family is sans}
}
\description{
This theme provides an figure layout that is
consistent with the style guidelines for the LOMU Annual Report
specifically for acoustic density plots on the lake.
}
\examples{
require(ggplot2)
ggplot(LOschaner, aes(lon, lat)) + geom_path()+ theme_hacmap()
}
\author{
Jeremy Holden
}
