% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescriptions.R
\docType{data}
\name{USGS_Species}
\alias{USGS_Species}
\title{USGS_Species}
\format{
A data frame
}
\usage{
USGS_Species
}
\description{
RVCAT reference table for SPECIES codes and COMMON_NAME
}
\keyword{datasets}
